alex-video-editor/auto_edit.py
"""
AUTO VIDEO EDITOR v1.0
Automated video editing with silence removal, captions, and music

Usage:
    python auto_edit.py input_video.mp4
    
Features:
    - Auto silence removal (40% shorter videos)
    - AI-generated captions (Whisper)
    - Background music with auto-ducking
    - Vertical 9:16 format conversion
    - Simple branding overlay
"""

import os
import sys
import yaml
from datetime import datetime
from moviepy import VideoFileClip, AudioFileClip, TextClip, CompositeVideoClip
from pydub import AudioSegment
from pydub.silence import detect_nonsilent
import numpy as np

# Whisper is optional (only needed if captions are enabled)
try:
    import whisper
    WHISPER_AVAILABLE = True
except ImportError:
    WHISPER_AVAILABLE = False
    print("Note: Whisper not installed. Captions will be disabled.")

class VideoEditor:
    def __init__(self, config_path="config.yaml"):
        """Initialize the video editor with config"""
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        print("=" * 60)
        print("  AUTO VIDEO EDITOR v1.0")
        print("=" * 60)
        print("")
        
    def remove_silence(self, audio_path, output_path):
        """Remove silence from audio track"""
        print(">> Removing silence...")
        
        # Load audio
        audio = AudioSegment.from_file(audio_path)
        
        # Detect non-silent parts
        nonsilent_ranges = detect_nonsilent(
            audio,
            min_silence_len=self.config['silence']['min_silence_ms'],
            silence_thresh=self.config['silence']['threshold_db']
        )
        
        if not nonsilent_ranges:
            print("   WARNING: No speech detected!")
            return audio_path
        
        # Add padding
        padding = self.config['silence']['padding_ms']
        padded_ranges = []
        for start, end in nonsilent_ranges:
            padded_start = max(0, start - padding)
            padded_end = min(len(audio), end + padding)
            padded_ranges.append((padded_start, padded_end))
        
        # Concatenate non-silent parts
        output_audio = AudioSegment.empty()
        for start, end in padded_ranges:
            output_audio += audio[start:end]
        
        # Export
        output_audio.export(output_path, format="mp3")
        
        original_duration = len(audio) / 1000
        new_duration = len(output_audio) / 1000
        reduction = ((original_duration - new_duration) / original_duration) * 100
        
        print(f"   Original: {original_duration:.1f}s")
        print(f"   Trimmed: {new_duration:.1f}s")
        print(f"   Reduced by: {reduction:.1f}%")
        
        return output_path
    
    def generate_captions(self, audio_path):
        """Generate captions using Whisper AI"""
        if not WHISPER_AVAILABLE:
            print("   WARNING: Whisper not available, skipping captions")
            return []
            
        print(">> Generating captions with Whisper AI...")
        
        # Load Whisper model (base is good balance of speed/accuracy)
        model = whisper.load_model("base")
        
        # Transcribe
        result = model.transcribe(audio_path, word_timestamps=True)
        
        captions = []
        for segment in result['segments']:
            captions.append({
                'start': segment['start'],
                'end': segment['end'],
                'text': segment['text'].strip()
            })
        
        print(f"   Generated {len(captions)} caption segments")
        return captions
    
    def create_caption_clips(self, captions, video_duration):
        """Create text overlay clips from captions"""
        print(">> Creating caption overlays...")
        
        text_clips = []
        cfg = self.config['captions']
        
        for caption in captions:
            # Create text clip
            txt = TextClip(
                text=caption['text'],
                font_size=cfg['font_size'],
                color=cfg['color'],
                font=cfg['font'],
                stroke_color=cfg['stroke_color'],
                stroke_width=cfg['stroke_width'],
                text_align='center',
                size=(1000, None)
            )
            
            # Set timing
            txt = txt.with_start(caption['start']).with_duration(caption['end'] - caption['start'])
            
            # Position
            if cfg['position'] == 'bottom':
                txt = txt.with_position(('center', 1600))
            else:
                txt = txt.with_position('center')
            
            text_clips.append(txt)
        
        return text_clips
    
    def add_music_with_ducking(self, video_clip, voice_audio_path):
        """Add background music with auto-ducking"""
        print(">> Adding background music...")
        
        music_file = self.config['music']['file']
        if not os.path.exists(music_file):
            print(f"   WARNING: Music file not found: {music_file}")
            return video_clip
        
        # Load music
        music = AudioFileClip(music_file)
        
        # Loop music to match video duration
        if music.duration < video_clip.duration:
            loops = int(video_clip.duration / music.duration) + 1
            music = music.loop(loops)
        
        # Trim to video duration
        music = music.subclipped(0, video_clip.duration)
        
        # Set volume
        music = music.with_volume_scaled(self.config['music']['volume'])
        
        # TODO: Implement ducking (Phase 1.5)
        # For now, just mix at constant volume
        
        # Mix audio
        final_audio = CompositeAudioClip([video_clip.audio, music])
        video_clip = video_clip.with_audio(final_audio)
        
        print("   Music added successfully")
        return video_clip
    
    def convert_to_vertical(self, video_clip):
        """Convert to 9:16 vertical format"""
        print(">> Converting to vertical format...")
        
        w, h = video_clip.size
        target_w = self.config['video']['resolution']['width']
        target_h = self.config['video']['resolution']['height']
        
        # If already vertical, just resize
        if h > w:
            return video_clip.resized((target_w, target_h))
        
        # Crop to vertical
        crop_w = int(h * 9 / 16)
        x_center = w // 2
        x1 = x_center - crop_w // 2
        
        video_clip = video_clip.cropped(x1=x1, x2=x1+crop_w)
        video_clip = video_clip.resized((target_w, target_h))
        
        print(f"   Converted to {target_w}x{target_h}")
        return video_clip
    
    def process_video(self, input_path):
        """Main processing pipeline"""
        print(f">> Processing: {os.path.basename(input_path)}")
        print("")
        
        # Create temp directory
        os.makedirs("temp", exist_ok=True)
        os.makedirs(self.config['output']['directory'], exist_ok=True)
        
        # Step 1: Load video
        print(">> Loading video...")
        video = VideoFileClip(input_path)
        print(f"   Duration: {video.duration:.1f}s")
        print(f"   Size: {video.w}x{video.h}")
        print("")
        
        # Step 2: Extract and process audio
        temp_audio = "temp/original_audio.mp3"
        video.audio.write_audiofile(temp_audio, logger=None)
        
        if self.config['silence']['enabled']:
            processed_audio = "temp/processed_audio.mp3"
            self.remove_silence(temp_audio, processed_audio)
        else:
            processed_audio = temp_audio
        
        print("")
        
        # Step 3: Generate captions
        if self.config['captions']['enabled']:
            captions = self.generate_captions(processed_audio)
        else:
            captions = []
        
        print("")
        
        # Step 4: Trim video to match processed audio
        if self.config['silence']['enabled']:
            processed_audio_clip = AudioFileClip(processed_audio)
            new_duration = processed_audio_clip.duration
            
            # Speed up video to match trimmed audio
            speed_factor = video.duration / new_duration
            video = video.with_speed_multiplier(speed_factor)
            video = video.with_audio(processed_audio_clip)
        
        # Step 5: Convert to vertical
        if self.config['video']['output_format'] == "9:16":
            video = self.convert_to_vertical(video)
        
        print("")
        
        # Step 6: Add captions
        if captions:
            caption_clips = self.create_caption_clips(captions, video.duration)
            video = CompositeVideoClip([video] + caption_clips)
        
        # Step 7: Add music
        if self.config['music']['enabled'] and os.path.exists(self.config['music']['file']):
            video = self.add_music_with_ducking(video, processed_audio)
        
        # Step 8: Trim to target duration
        target_duration = self.config['video']['target_duration']
        if video.duration > target_duration:
            print(f">> Trimming to {target_duration}s...")
            video = video.subclipped(0, target_duration)
        
        # Step 9: Export
        output_filename = self.config['output']['filename_template'].format(
            name=os.path.splitext(os.path.basename(input_path))[0],
            date=datetime.now().strftime("%Y%m%d_%H%M%S")
        )
        output_path = os.path.join(self.config['output']['directory'], output_filename)
        
        print("")
        print(">> Exporting final video...")
        print(f"   Output: {output_path}")
        print("   This may take 2-5 minutes...")
        print("")
        
        video.write_videofile(
            output_path,
            fps=self.config['video']['fps'],
            codec='libx264',
            audio_codec='aac',
            preset='medium',
            threads=4,
            logger='bar'
        )
        
        # Cleanup
        if not self.config['processing']['keep_temp_files']:
            import shutil
            shutil.rmtree("temp")
        
        print("")
        print("=" * 60)
        print("  SUCCESS!")
        print(f"  Video saved to: {output_path}")
        print("=" * 60)
        print("")
        print(">> Ready to post to Instagram/TikTok!")
        print("")
        
        return output_path

def main():
    if len(sys.argv) < 2:
        print("Usage: python auto_edit.py input_video.mp4")
        sys.exit(1)
    
    input_video = sys.argv[1]
    
    if not os.path.exists(input_video):
        print(f"Error: File not found: {input_video}")
        sys.exit(1)
    
    editor = VideoEditor()
    editor.process_video(input_video)

if __name__ == "__main__":
    main()
