alex-video-editor/requirements.txt
# Auto Video Editor v1.0 - Requirements
# Install with: pip install -r requirements.txt

# Core video processing
moviepy==2.1.2
pydub==0.25.1

# AI transcription
openai-whisper==20231117

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
numpy==2.2.6
pillow>=9.2.0,<11.0  # Compatible with MoviePy
tqdm==4.67.1
pyyaml==6.0.1
